Remove GWT detection for now
There is no way of accessing iframe.contentWindow on 3rd party domains
without triggering a warning in the console. I'm not familiar with GWT,
so I'm not sure if this iframe method is a requirement or not. Since
this is blocking a release though, we can address this as a minor
version fix later.