Harden React detection
This backs out the change done in #96 that was causing a false positive for React.  `window.__REACT_DEVTOOLS_GLOBAL_HOOK__` will return `true` for anyone that has React Devtools installed, which isn't necessarily representative of what is running on a given page.  In addition, this adds a check for a `#react-root` element and updates the methods checked to verify React existence.